src/raspberry/requirements.txt
# Note: Some packages are installed both via apt and pip
# to ensure compatibility. The following packages are also
# installed via apt in setup_raspberry_complete.sh:
# - python3-opencv (equivalent to opencv-python)
# - python3-picamera2
# - python3-libcamera (system dependency for picamera2)

# Communication
pyserial==3.5        # Serial communication with Arduino
python-osc>=1.8.1    # OSC communication

# Image and video processing
opencv-python>=4.8.0  # OpenCV for video capture
numpy>=1.24.0        # Numerical calculations for OpenCV
picamera2>=0.3.9     # Library for Raspberry Pi Camera on Pi 5

# Raspberry Pi Hardware
RPi.GPIO>=0.7.0      # Raspberry Pi GPIO control

# Environment
python-dotenv>=0.19.0  # Environment variables management

# System interaction
python-prctl>=1.8.1   # Process control functionality (needed by picamera2)